pyproject.toml
[tool.poetry]
name = "myfastapi"
version = "0.1.0"
description = "企业级 FastAPI + MySQL 后台管理系统"
authors = ["JiangJiangYo <your_email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13.8"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.0"
pydantic = "^2.7.0"
mysqlclient = "^2.2.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"